requirements.txt
gradio == 3.27.0
openai == 0.27.4
python-dotenv == 1.0.0
langchain
chromadb == 0.3.21
tiktoken == 0.3.3
pypdf == 3.8.1
pypdf2
faiss-cpu